Applied changes according to review comments.
