+++ src/Lecture_1/Homework_1/Task_3.java
@@ -1,4 +1,4 @@
-package Lecture_1.Homework;
+package Lecture_1.Homework_1;
 
 import java.util.Scanner;
